Improve timer for settings service
The settings service is not available at the beginning of late_start because services are launched in this mode. Waiting 10s will be enough in most cases. But this is a race condition and it can be lost sometimes. In order to improve the chances without a waiting time too long for the user, a watcher is added to get when settings service is ready. The watcher timeout after 2min.